% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_me.R
\name{add_me}
\alias{add_me}
\title{Add Author Information into DESCRIPTION}
\usage{
add_me(is_paste = FALSE)
}
\arguments{
\item{is_paste}{Logical. By default, \code{FALSE}}
}
\description{
Add Author Information into DESCRIPTION
}
