% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zenodo_citation.R
\name{add_zenodo_citation}
\alias{add_zenodo_citation}
\title{Add the Zenodo citation}
\usage{
add_zenodo_citation(readme_path = here::here("../pyks/README.Rmd"))
}
\arguments{
\item{readme_path}{here::here object}
}
\description{
Add the Zenodo citation
}
