% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_use_package.R
\name{map_use_package}
\alias{map_use_package}
\title{Use package multiply}
\usage{
map_use_package(input)
}
\arguments{
\item{input}{The list of package names.}
}
\value{
None
}
\description{
This function is similar to the function \code{use_package} from the package \code{usethis},
but its input can be a list, like \code{c('clipr', 'glue', 'rstudioapi', 'rvest', 'usethis', 'xml2')}
When the package developers receive the error information using \code{devtools::check()},
\code{Namespace dependencies not required: ...}, use this.
}
\examples{
\dontrun{map_use_package(c('dplyr', 'glue', 'purrr', 'stringr'))}
}
