% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_disclaimer.R
\name{add_disclaimer}
\alias{add_disclaimer}
\title{Add disclaimer}
\usage{
add_disclaimer(name = "Jiaxiang Li", license_name = FALSE,
  github_id = "JiaxiangBU")
}
\arguments{
\item{name}{Author full name.}

\item{license_name}{License name.}

\item{github_id}{Author GitHub Id}
}
\value{
The text contains the disclaimer with author name and license name.
}
\description{
Add disclaimer
}
