% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_desc.R
\name{create_desc}
\alias{create_desc}
\title{Create the description file for R project}
\usage{
create_desc(set_name = TRUE)
}
\arguments{
\item{set_name}{Logical. By default, \code{TRUE}}
}
\value{
None
}
\description{
Create the description file for R project
}
